server/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
openai
python-dotenv
